Replace squashfuse binary
The new binary is squashfuse_ll built in an Ubuntu 18.04 Docker. It has xz and lz4 statically linked as they are outside the [AppImage exclude list](github.com/AppImage/pkg2appimage/blob/master/excludelist). I wanted to add ZSTD, but adding it statically GREATLY bloats the size of the binary (from 200kB to about 500kB). The new binary is also named with the architecture, because I also plan on building for ARM and possibly others