Update to latest Zig, some cleanup, mounting works and FUSE improvements
When mounting the example with the `-s` flag, mounting appears to work
as intended. `AppRun` can now be launched from the AppImages I've tested
on. FUSE error messages have now been converted into an enum for all
functions that permit it, which makes the code much more readable
without a bunch of `@intToEnum`s. I've also done a bit of cleaning up on
the main example, which I'll split off into a clone of the squashfuse_hl
binary along with removing some of the bad hacks I threw together to get
stuff sorta working initially.